update clickable examples
