+++ packages/wonder-blocks-clickable/generated-snapshot.test.js
@@ -15,6 +15,45 @@ describe("wonder-blocks-clickable", () => {
         const {StyleSheet} = require("aphrodite");
         const {View} = require("@khanacademy/wonder-blocks-core");
         const {MemoryRouter, Route, Switch} = require("react-router-dom");
+        const Color = require("@khanacademy/wonder-blocks-color").default;
+        const {Body} = require("@khanacademy/wonder-blocks-typography");
+
+        const styles = StyleSheet.create({
+            focused: {
+                border: "none",
+            },
+            hovered: {
+                textDecoration: "underline",
+            },
+            pressed: {
+                color: Color.teal,
+            },
+        });
+
+        const example = (
+            <View>
+                <Clickable onClick={() => alert("You clicked some text!")}>
+                    {(eventState) => (
+                        <Body
+                            style={[
+                                eventState.focused && styles.focused,
+                                eventState.hovered && styles.hovered,
+                                eventState.pressed && styles.pressed,
+                            ]}
+                        >
+                            This text is clickable!
+                        </Body>
+                    )}
+                </Clickable>
+            </View>
+        );
+        const tree = renderer.create(example).toJSON();
+        expect(tree).toMatchSnapshot();
+    });
+    it("example 2", () => {
+        const {StyleSheet} = require("aphrodite");
+        const {View} = require("@khanacademy/wonder-blocks-core");
+        const {MemoryRouter, Route, Switch} = require("react-router-dom");
 
         const styles = StyleSheet.create({
             row: {
