+++ packages/wonder-blocks-core/generated-snapshot.test.js
@@ -9,7 +9,6 @@ import renderer from "react-test-renderer";
 // Mock react-dom as jest doesn't like findDOMNode.
 jest.mock("react-dom");
 import ClickableBehavior from "./components/clickable-behavior.js";
-import Clickable from "./components/clickable.js";
 import IDProvider from "./components/id-provider.js";
 import Text from "./components/text.js";
 import UniqueIDProvider from "./components/unique-id-provider.js";
@@ -18,77 +17,6 @@ import WithSSRPlaceholder from "./components/with-ssr-placeholder.js";
 
 describe("wonder-blocks-core", () => {
     it("example 1", () => {
-        const {
-            LabelLarge,
-            LabelMedium,
-        } = require("@khanacademy/wonder-blocks-typography");
-        const Icon = require("@khanacademy/wonder-blocks-icon").default;
-        const {icons} = require("@khanacademy/wonder-blocks-icon");
-        const {Spring, Strut} = require("@khanacademy/wonder-blocks-layout");
-        const {StyleSheet} = require("aphrodite");
-
-        const styles = StyleSheet.create({
-            container: {
-                display: "flex",
-                flexDirection: "row",
-                padding: 16,
-                border: "solid 1px black",
-                borderRadius: 4,
-                width: "100%",
-            },
-            hovered: {
-                backgroundColor: "lightgreen",
-                cursor: "pointer",
-            },
-            pressed: {
-                backgroundColor: "green",
-                color: "white",
-            },
-            focused: {
-                "::before": {
-                    content: "''",
-                    position: "absolute",
-                    left: -8,
-                    top: -8,
-                    right: -8,
-                    bottom: -8,
-                    borderRadius: 10,
-                    border: "dotted 4px black",
-                },
-            },
-        });
-
-        const example = (
-            <View>
-                <Clickable>
-                    {(state) => {
-                        console.log(state);
-                        return (
-                            <View
-                                style={[
-                                    styles.container,
-                                    state.hovered && styles.hovered,
-                                    state.pressed && styles.pressed,
-                                    state.focused && styles.focused,
-                                ]}
-                            >
-                                <View style={{alignItems: "flex-start"}}>
-                                    <LabelLarge>Hello, world</LabelLarge>
-                                    <Strut size={8} />
-                                    <LabelMedium>Lorem Ipsum Dolor</LabelMedium>
-                                </View>
-                                <Spring />
-                                <Icon size="large" icon={icons.search} />
-                            </View>
-                        );
-                    }}
-                </Clickable>
-            </View>
-        );
-        const tree = renderer.create(example).toJSON();
-        expect(tree).toMatchSnapshot();
-    });
-    it("example 2", () => {
         const example = (
             <View>
                 <IDProvider scope="field">
@@ -104,7 +32,7 @@ describe("wonder-blocks-core", () => {
         const tree = renderer.create(example).toJSON();
         expect(tree).toMatchSnapshot();
     });
-    it("example 3", () => {
+    it("example 2", () => {
         const example = (
             <View>
                 <IDProvider scope="field" id="some-user-id">
@@ -120,7 +48,7 @@ describe("wonder-blocks-core", () => {
         const tree = renderer.create(example).toJSON();
         expect(tree).toMatchSnapshot();
     });
-    it("example 4", () => {
+    it("example 3", () => {
         const {StyleSheet} = require("aphrodite");
 
         const styles = StyleSheet.create({
@@ -150,7 +78,7 @@ describe("wonder-blocks-core", () => {
         const tree = renderer.create(example).toJSON();
         expect(tree).toMatchSnapshot();
     });
-    it("example 5", () => {
+    it("example 4", () => {
         const example = (
             <View>
                 <View onClick={() => alert("Clicked!")}>Click me!</View>
@@ -163,7 +91,7 @@ describe("wonder-blocks-core", () => {
         const tree = renderer.create(example).toJSON();
         expect(tree).toMatchSnapshot();
     });
-    it("example 6", () => {
+    it("example 5", () => {
         const example = (
             <View>
                 <View testId="foo">Foo</View>
@@ -173,7 +101,7 @@ describe("wonder-blocks-core", () => {
         const tree = renderer.create(example).toJSON();
         expect(tree).toMatchSnapshot();
     });
-    it("example 7", () => {
+    it("example 6", () => {
         const {
             Body,
             HeadingSmall,
@@ -224,7 +152,7 @@ describe("wonder-blocks-core", () => {
         const tree = renderer.create(example).toJSON();
         expect(tree).toMatchSnapshot();
     });
-    it("example 8", () => {
+    it("example 7", () => {
         const {
             Body,
             BodyMonospace,
@@ -267,7 +195,7 @@ describe("wonder-blocks-core", () => {
         const tree = renderer.create(example).toJSON();
         expect(tree).toMatchSnapshot();
     });
-    it("example 9", () => {
+    it("example 8", () => {
         const {
             Body,
             HeadingSmall,
@@ -298,7 +226,7 @@ describe("wonder-blocks-core", () => {
         const tree = renderer.create(example).toJSON();
         expect(tree).toMatchSnapshot();
     });
-    it("example 10", () => {
+    it("example 9", () => {
         const {
             BodyMonospace,
         } = require("@khanacademy/wonder-blocks-typography");
@@ -322,7 +250,7 @@ describe("wonder-blocks-core", () => {
         const tree = renderer.create(example).toJSON();
         expect(tree).toMatchSnapshot();
     });
-    it("example 11", () => {
+    it("example 10", () => {
         const {StyleSheet} = require("aphrodite");
 
         const styles = StyleSheet.create({
@@ -352,7 +280,7 @@ describe("wonder-blocks-core", () => {
         const tree = renderer.create(example).toJSON();
         expect(tree).toMatchSnapshot();
     });
-    it("example 12", () => {
+    it("example 11", () => {
         const example = (
             <View>
                 <View onClick={() => alert("Clicked!")}>Click me!</View>
@@ -365,7 +293,7 @@ describe("wonder-blocks-core", () => {
         const tree = renderer.create(example).toJSON();
         expect(tree).toMatchSnapshot();
     });
-    it("example 13", () => {
+    it("example 12", () => {
         const example = (
             <WithSSRPlaceholder
                 placeholder={() => (
@@ -386,7 +314,7 @@ describe("wonder-blocks-core", () => {
         const tree = renderer.create(example).toJSON();
         expect(tree).toMatchSnapshot();
     });
-    it("example 14", () => {
+    it("example 13", () => {
         const example = (
             <WithSSRPlaceholder placeholder={null}>
                 {() => (
@@ -400,7 +328,7 @@ describe("wonder-blocks-core", () => {
         const tree = renderer.create(example).toJSON();
         expect(tree).toMatchSnapshot();
     });
-    it("example 15", () => {
+    it("example 14", () => {
         const {
             Body,
             BodyMonospace,
@@ -479,7 +407,7 @@ describe("wonder-blocks-core", () => {
         const tree = renderer.create(example).toJSON();
         expect(tree).toMatchSnapshot();
     });
-    it("example 16", () => {
+    it("example 15", () => {
         const {
             Body,
             BodyMonospace,
